requirements.txt
python-telegram-bot[job_queue]
SQLAlchemy
anki-sm-2
